Fix leak on ABFRealmTableViewController dismissal
Here is a demonstration of the leak: https://github.com/0xced/ABFRealmTableViewController/tree/memory-leak-demo

#### No leak
1. Push a new `MainTableViewController` (using the fast forward button in the top left corner)
2. Go back to root view controller without doing anything
3. Observe that the pushed `MainTableViewController` is released as proved by the log `-[MainTableViewController dealloc] 0x7fe613d0a410`

#### MainTableViewController leaked
1. Push a new `MainTableViewController` (using the fast forward button in the top left corner)
2. Create and/or delete a task entry
3. Go back to root view controller
4. Observe that the pushed `MainTableViewController` is **not** released (no dealloc log plus you can see it with Xcode's memory graph)

Resetting the fetched results controller in `viewDidDisappear:` solves the leak and the pushed `MainTableViewController` will be released. I have not fully understood why resetting the fetched results controller solves the leak, this is certainly somehow related to the change notifications.

Note that the leak does not happen if the project is compiled with the Release configuration instead of the Debug configuration, even without the fetched results controller. But this is preferable to have a solution that works for both Debug and Release configurations.